DPII instructions fully refactored to share logic with DPIR instructions. Wide move logic fixed.

See merge request lab2526_summer/armv8_58!15

+++ src/emulator/decode.h
@@ -66,7 +66,7 @@ typedef enum OpType {
 
 //
 typedef struct ImmediateArithmeticInstruction {
-    int imm12;
+    uint16 imm12;
     int rd;
     int rn;
     bool sf;
